+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.1] - 2023-01-22
+
+### Changed
+
+- quote uses SecuritySymbol instead of separate exchange and symbol parameters.
+
+### Fixed
+
+- better symbol parsing on prune
+
 ## [1.5.0] - 2023-01-22
 
 ### Added